chore: seperated PaginationCarousel into its own component
