+++ src/components/layout/Main.tsx
@@ -1,21 +1,24 @@
+'use client';
+
 import { type ReactNode } from 'react';
 
 import { cx } from '@/lib/utils';
 
-type Props = {
+type MainProps = {
   children?: ReactNode;
   mainClassName?: string;
   className?: string;
 };
 
-function Main({ children, mainClassName, className }: Props) {
+function Main({ children, mainClassName, className }: MainProps) {
   return (
     <main className={cx('flex w-full flex-grow justify-center', mainClassName)}>
       <div
         className={cx(
-          'h-full w-full max-w-screen-2xl px-4 sm:px-11 md:px-16 lg:px-24',
+          'h-full w-full max-w-screen-2xl px-4 duration-1000 animate-in fade-in max-md:slide-in-from-left-8 sm:px-11 md:px-16 md:slide-in-from-top-8 lg:px-24',
           className,
         )}
+        id='main'
       >
         {children}
       </div>
